Implemented automated releases to Clojars
Since our goal is to release more often I implemented a first draft
to automate the releases. Releases are a very individual thing and
I tried to cover datahikes previous manual process as far as
possible. This means that a commit to development creates a release
if there is a SNAPSHOT-version specified in project.clj and a
commit to master creates a release if there is no SNAPSHOT-version
specified in project.clj.

Besides that there is some streamlining of the steps happening in
the usual testing and building procedure. Furthermore I introduced
the new workflow-feature of CircleCI and now the steps are able to
run concurrently. This seems to be quite a bit faster.

- Refers https://github.com/replikativ/datahike/issues/167
- TODO automate tagging
- TODO automate image building or use different one